// apps/northwind/src/app/store/customer/types.ts
import { Customer, ListMetadata } from '@api-interfaces';

export interface ICustomer {
	data: Customer[];
	metadata: ListMetadata;
	isLoading: boolean;
	error: string;
	customerOnEdition: Customer,
}

export enum CUSTOMERS {
	RESET_DATA = 'CUSTOMERS_RESET_DATA',
	RESET_EDIT_DATA = 'CUSTOMERS_RESET_EDIT_DATA',

	INIT = 'CUSTOMERS_INIT',
	SUCCESS = 'CUSTOMERS_SUCCESS',
	ERROR = 'CUSTOMERS_ERROR',
	DELETED = 'CUSTOMERS_DELETED',
	TERMINATE = 'CUSTOMERS_TERMINATE',

	CREATE_INIT = 'CUSTOMERS_CREATE_INIT',
	CREATE_SUCCESS = 'CUSTOMERS_CREATE_SUCCESS',
	CREATE_ERROR = 'CUSTOMERS_CREATE_ERROR',

	OPEN_UPDATE = 'CUSTOMERS_OPEN_UPDATE',
	UPDATE_INIT = 'CUSTOMERS_UPDATE_INIT',
	UPDATE_SUCCESS = 'CUSTOMERS_UPDATE_SUCCESS',
	UPDATE_ERROR = 'CUSTOMERS_UPDATE_ERROR',
}